% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_dob_process.R
\name{age_dob_process}
\alias{age_dob_process}
\title{Age and DOB processing}
\usage{
age_dob_process(
  data.var,
  age.var,
  dob.var,
  date.var,
  lim1,
  lim2,
  lim3 = NULL,
  lim4 = NULL,
  lim5 = NULL
)
}
\arguments{
\item{data.var}{Data table}

\item{age.var}{Variable containing age in years.}

\item{dob.var}{Variable containing date of birth}

\item{date.var}{Variable containing enrolment/consent date}

\item{lim1}{Age limit for categorization}

\item{lim2}{Age limit for categorization}

\item{lim3}{Age limit for categorization}

\item{lim4}{Age limit for categorization}

\item{lim5}{Age limit for categorization}
}
\value{
Returns the data table with a calculated age column and categorized age column added.
}
\description{
The function first calculates the age based on the present age variable and if unavailable date of birth and enrollment/consent date, end then runs the age_groups function to divide into different categories.
}
\examples{
library(dplyr)
library(lubridate)


my_data <- data.frame(
  ID = c(1, 2, 3, 4, 5),
  age = c(25, NA, 30, NA, NA),
  dob = c("1980-01-15", "1995-03-22", "1975-08-10", "1988-11-05", "2000-09-30"),
  date_enrollment = c("2023-05-10", "2022-12-01", "2022-02-18", "2023-02-05", "2021-07-20")
)

# Display the data
print(my_data)

# Apply the age_dob_process function to the data
age_dob_process(data.var=my_data, age.var = "age", dob.var = "dob", date.var = "date_enrollment", lim1 = 30, lim2 = 40)


}
