% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_result.R
\name{categorize_result}
\alias{categorize_result}
\title{categorize_result}
\usage{
categorize_result(dataframe, variable, thresholds)
}
\arguments{
\item{dataframe}{The data table contaiing test results}

\item{variable}{A continuous variable}

\item{thresholds}{The thresholds to be applied}
}
\value{
Data table with the categorization added
}
\description{
Categorize a subject as "Positive" or "Negative", based on a variable of interest and a specified threshold. The values are bigger or equal to the threshold are categorized as "Positive", others "Negative".
A list of thresholds can be provided.
}
\examples{

df <- data.frame(results = c(sample(1:100, 50, replace=TRUE)))
categorize_result(dataframe = df, variable = "results", thresholds = 50)
}
