% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_result_pcr.R
\name{categorize_result_PCR}
\alias{categorize_result_PCR}
\title{categorize_result_PCR}
\usage{
categorize_result_PCR(dataframe, variable, thresholds)
}
\arguments{
\item{dataframe}{The data table containing test results}

\item{variable}{A continuous variable - PCR Ct data}

\item{thresholds}{The thresholds to be applied}
}
\value{
Data table with category added
}
\description{
Same as categorize_result but the values are smaller than the threshold are categorized as "Positive", others "Negative".
}
