% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_calc.R
\name{CI_Calc}
\alias{CI_Calc}
\title{Confidence Interval Calculator}
\usage{
CI_Calc(alpha, proportion, n)
}
\arguments{
\item{alpha}{significance level}

\item{proportion}{the proportion}

\item{n}{sample size}
}
\value{

}
\description{
Calculates the confidence interval of a given proportion.
}
\examples{
CI_Calc(alpha=0.05, proportion=0.8, n= 321)
}
