% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen_agreement.R
\name{cohen_agreement}
\alias{cohen_agreement}
\title{Cohen agreement}
\usage{
cohen_agreement(data_var, var1, var2)
}
\arguments{
\item{data_var}{data table}

\item{var1}{The test with "Positive" and "Negative"results.}

\item{var2}{The comparator with "Positive" and "Negative"results.}
}
\value{
Data table
}
\description{
Calculates Cohen's kappa and generates a table including number of positive and negative agreement, as well as positive, negative,  overall percent agreements, and the Kappa value.
}
\examples{

df <- data.frame(Test1 = c(rep("Positive", 20), rep("Negative", 20)), Test2 = c(rep("Positive", 30), rep("Negative", 10)))
cohen_agreement(data_var = df, var1 = "Test1", var2 = "Test2")

}
