% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_for_roc.R
\name{composite_for_roc}
\alias{composite_for_roc}
\title{composite_for_roc}
\usage{
composite_for_roc(ref_1, ref_2, test_refs, data_frame)
}
\arguments{
\item{ref_1}{Reference 1}

\item{ref_2}{Reference 2}

\item{test_refs}{Reference categorized based on multiple thresholds}

\item{data_frame}{Data table}
}
\value{
Data table containing all possible composite references based on the thresholds applied
}
\description{
Define multiple composite references to make a ROC style analysis but on the references.
}
