% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_reference.R
\name{composite_reference}
\alias{composite_reference}
\title{composite_reference}
\usage{
composite_reference(ref_1, ref_2, ref_3 = NULL, data_frame)
}
\arguments{
\item{ref_1}{Test1 containing "Positive or "Negative" Results}

\item{ref_2}{Test2 containing "Positive or "Negative" Results}

\item{ref_3}{Test3 containing "Positive or "Negative" Results}

\item{data_frame}{Data table containing the 3 tests}
}
\value{
Data table with composite reference added
}
\description{
Generates a composite reference based on 3 tests. When any of the tests is positive, subject is classified as positive.
Works with 2 or 3 reference tests.
}
\examples{
df <- data.frame(T1 = c(rep("Positive", 10), rep("Negative", 10), rep("Negative", 10)), T2 = c(rep("Positive", 10), rep("Positive", 10), rep("Negative", 10)))
composite_reference(ref_1 = "T1", ref_2 = "T2", ref_3 = NULL, data_frame = df)
}
