% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_yield.R
\name{dx_yield}
\alias{dx_yield}
\title{Diagnostic Yield
Calculates Diagnostic yield: # Positives per Index Test/ # Positive by Reference Test}
\usage{
dx_yield(data_arg, ref_arg, index_arg, conf.level = 0.95)
}
\arguments{
\item{data_arg}{The data frame}

\item{ref_arg}{Defined reference column}

\item{index_arg}{The index test results column}

\item{conf.level}{Level of confidence}
}
\value{
Diagnostic yield with confidence intervals
}
\description{
Diagnostic Yield
Calculates Diagnostic yield: # Positives per Index Test/ # Positive by Reference Test
}
\examples{


df_ex <- data.frame(REF = c(rep('Positive', 50), rep('Negative', 50)), INDEX = c(rep('Positive', 3), rep('Negative', 78), rep('Positive', 19)))

dx_yield(data_arg = df_ex, ref_arg = "REF", index_arg = "INDEX", conf.level = 0.95)

}
