% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finddataanalysis-package.R
\docType{package}
\name{finddataanalysis-package}
\alias{finddataanalysis}
\alias{finddataanalysis-package}
\title{finddataanalysis: FIND Data Analysis Functions}
\description{
Includes generic functions that are frequently used for data analyses at FIND.
}
\author{
\strong{Maintainer}: Berra Erkosar \email{berra.erkosar@finddx.org}

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
