% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_means_concordence.R
\name{group_means_concordence}
\alias{group_means_concordence}
\title{Group Means - Supplement for Concordance Analysis}
\usage{
group_means_concordence(data_var, mean_var, var1, var2, var3)
}
\arguments{
\item{data_var}{The data table}

\item{mean_var}{The continuous variable from which the mean will be calculated}

\item{var1}{A bifactor variable with "Positive" and "Negative" results.}

\item{var2}{A bifactor variable with "Positive" and "Negative" results.}

\item{var3}{A bifactor variable with "Positive" and "Negative" results.}
}
\description{
Calculate group means of a continuous variable of interest in all possible combinations if agreements between three reference columns
}
\examples{

df <- data.frame(Test1 = c(rep("Positive", 20), rep("Negative", 20)), Test2 = c(rep("Negative", 5), rep("Positive", 5), rep("Negative", 30)), Test3 = c(rep("Positive", 10), rep("Negative", 25), rep("Positive", 5)), MyVariable = sample(1:100, 40))


group_means_concordence(data_var = df, mean_var = "MyVariable", var1 = "Test1", var2 = "Test2", var3 = "Test3")

}
