% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sen_spe_forest.R
\name{multi_sen_spe_forest}
\alias{multi_sen_spe_forest}
\title{multi_sen_spe_forest for forest plots}
\usage{
multi_sen_spe_forest(
  data_var,
  list_index,
  ref,
  conf.level = 0.95,
  index_names = NULL
)
}
\arguments{
\item{data_var}{Data table containing test results}

\item{list_index}{A list of index test with "Negative" and "Positive" results}

\item{ref}{The reference test with "Negative" and "Positive" results}

\item{conf.level}{The confidence level, 1-alpha – default 95\%}

\item{index_names}{An optional list containing names of the index tests}
}
\value{
A data table with TP/FP/TN/FN, sensitivity, specificity and confidence intervals, to be fed into forestplot function.
}
\description{
Calculates and displays sensitivity and specificity with confidence intervals from a data frame for multiple index tests and the reference test. To be used for displaying the result in a table.
}
\examples{

df <- data.frame(Index1 = c(rep("Positive", 20), rep("Negative", 20)), Index2 = c(rep("Positive", 5), rep("Negative", 35)), Index3 = c(rep("Positive", 9), rep("Negative", 31)), Reference = c(rep("Positive", 10), rep("Negative", 30)))

multi_sen_spe_forest(data_var = df, list_index = c("Index1", "Index2", "Index3"), ref = "Reference", conf.level = 0.95, index_names = c("ate","FNDH", "DXCb"))
}
