% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sen_spe_out_forest.R
\name{multi_sen_spe_out_forest}
\alias{multi_sen_spe_out_forest}
\title{multi_sen_spe with forest plot output}
\usage{
multi_sen_spe_out_forest(
  data_var,
  list_index,
  ref,
  conf.level = 0.95,
  index_names = NULL,
  labels = "Tests"
)
}
\arguments{
\item{data_var}{Data table containing tests results}

\item{list_index}{A list of index tests with "Negative" and "Positive" results}

\item{ref}{The reference test with "Negative" and "Positive" results}

\item{conf.level}{The confidence level, 1-alpha – default 95\%}

\item{index_names}{An optional list containing names of the index tests}

\item{labels}{An optional text for the axis label}
}
\value{
Interactive forest plots in a list
}
\description{
Executes multi_sen_spe function and creates forest plots.
}
\examples{
df <- data.frame(Index1 = c(rep("Positive", 20), rep("Negative", 20)), Index2 = c(rep("Positive", 5), rep("Negative", 35)), Index3 = c(rep("Positive", 9), rep("Negative", 31)), Reference = c(rep("Positive", 10), rep("Negative", 30)))
plots <- multi_sen_spe_out_forest(data_var = df, list_index = c("Index1", "Index2", "Index3"), ref = "Reference", conf.level = 0.95, labels = "Index Tests", index_names = c("a", "b", "c"))
plots[[1]]
plots[[2]]
}
