% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table.R
\name{nice_table}
\alias{nice_table}
\title{nice_table}
\usage{
nice_table(data_var, group_name)
}
\arguments{
\item{data_var}{data frame}

\item{group_name}{Name of the grouping variable used in subgroups}
}
\value{
A nicely formatted DT table
}
\description{
To create a well formatted table from the basic data frame version of the performance characteristic table generated using dplyr for subgroup analyses, where the data is not formatted in the long format necessarily.
}
\examples{
library(dplyr)
df <- data.frame(Index1 = c(rep("Positive", 20), rep("Negative", 20)), Index2 = c(rep("Positive", 5), rep("Negative", 35)), Index3 = c(rep("Positive", 9), rep("Negative", 31)), Reference = c(rep("Positive", 10), rep("Negative", 30)), Sex = rep(c("M","F", "F", "M"), 10))

sex_ana <- df \%>\% group_by(Sex) \%>\% group_modify(~ performance_eval_auto(
  data_var = (.x),
  list_index =c("Index1", "Index2"),
  ref = "Reference",
  conf.level = 0.95,
  index_names = NULL,
  labels = "Test",
  forest_plot = FALSE, #!!!
  table_output = FALSE, #!!!
  file_name = "performance_eval",
  data_long = FALSE, #!!!
  group_var = NULL
))
nice_table(sex_ana, "Sex")
}
