% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_spe.R
\name{sens_spe}
\alias{sens_spe}
\title{Sensitivity and specificity (Table Display)}
\usage{
sens_spe(data_var, index, ref, conf.level = 0.95)
}
\arguments{
\item{data_var}{Data table containing test results}

\item{index}{The index test with "Negative" and "Positive" results}

\item{ref}{The reference test with "Negative" and "Positive" results}

\item{conf.level}{The confidence level, 1-alpha -- default 95\%}
}
\value{
A data table with TP/FP/TN/FN, sensitivity, specificity and confidence intervals
}
\description{
Calculate sensitivity and specificity with confidence intervals from a data frame that contains the results of the index and the reference test.
To be used for displaying the result in a table.
}
\examples{
df <- data.frame(Index = c(rep("Positive", 20), rep("Negative", 20)), Reference = c(rep("Positive", 10), rep("Negative", 30)))
sens_spe(data_var = df, index = "Index", ref = "Reference", conf.level = 0.95)
}
