% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_spe_for_forest.R
\name{sens_spe_for_forest}
\alias{sens_spe_for_forest}
\title{Sensitivity and specificity (for forest plots)}
\usage{
sens_spe_for_forest(data_var, index, ref, conf.level = 0.95)
}
\arguments{
\item{data_var}{Data table containing test results}

\item{index}{The index test with "Negative" and "Positive" results}

\item{ref}{The reference test with "Negative" and "Positive" results}

\item{conf.level}{The confidence level, 1-alpha -- default 95\%}
}
\value{
A data table with TP/FP/TN/FN, sensitivity, specificity and confidence intervals
}
\description{
Calculate sensitivity and specificity with confidence intervals from a data frame that contains the results of the index and the reference test.
To be used for displaying the result in a forest plot.
}
\examples{
library(forestplot)
df <- data.frame(Index = c(rep("Positive", 20), rep("Negative", 20)), Reference = c(rep("Positive", 30), rep("Negative", 10)))
forest_data <- cbind(Test = "Index", sens_spe_for_forest(data_var = df, index = "Index", ref = "Reference", conf.level = 0.95))
forest_data \%>\% forestplot(mean = Sensitivity, 
                          lower = SensLower, 
                          upper = SensUpper, 
                          labeltext = Test,
                          title = "Sensitivity",
                          zero = NA,
                          xticks = c( 0, 20, 40, 60, 80, 100),
                          txt_gp = fpTxtGp(ticks=gpar(cex=1)),
                          boxsize = .1
                          )
}
