% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_groups.R
\name{temperature_groups}
\alias{temperature_groups}
\title{Regroup axillary temperature values}
\usage{
temperature_groups(data.var, temperature)
}
\arguments{
\item{data.var}{Data table}

\item{temperature}{Variable containing body temperature in Celsius}
}
\value{
Creates a new column frame with body temperature ranges
}
\description{
Creates a new categorical variable based on the body temperature.
}
\examples{

df = data.frame(ID = c(1:20), Temperature = runif(20, 36, 42))
temperature_groups(data.var = df, temperature = "Temperature")
}
