% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_groups.R
\name{age_groups}
\alias{age_groups}
\title{Age groups}
\usage{
age_groups(data.var, age)
}
\arguments{
\item{data.var}{Data table}

\item{temperature}{Variable containing age in years.}
}
\value{

}
\description{
Creates a new categorical variable based on age.
}
\examples{
df = data.frame(ID = c(1:20), AGE = runif(20, 18, 86))
age_groups(data.var = df, age = "AGE")
}
