% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_reference_majority.R
\name{composite_reference_majority}
\alias{composite_reference_majority}
\title{Composite Reference Majority}
\usage{
composite_reference_majority(ref_1, ref_2, ref_3, data_frame)
}
\arguments{
\item{ref_1}{Test1 containing "Positive or "Negative" Results}

\item{ref_2}{Test2 containing "Positive or "Negative" Results}

\item{ref_3}{Test3 containing "Positive or "Negative" Results}

\item{data_frame}{Data table containing the 3 tests}
}
\value{
Data table with composite reference added
}
\description{
The function reads the results of 3 reference test results and takes the majority as the final result.
}
