% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_calculate.R
\name{ROC_calculate}
\alias{ROC_calculate}
\title{Title}
\usage{
ROC_calculate(data_var, index, ref, conf.level = 0.95, var_name_suffix_cut)
}
\arguments{
\item{data_var}{data table}

\item{index}{index test column containing results: "Positive" or "Negative"}

\item{ref}{a list containing names of the reference test columns}

\item{conf.level}{confidence level}

\item{var_name_suffix_cut}{the suffix based on which the reference test column name to be cut. This will result in suffix + threshold}
}
\description{
Calculates sensitivity and false positivity rate based on the multiple reference columns defined by the thresholds
}
